% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_page_header.R
\name{add_page_heading}
\alias{add_page_heading}
\title{clinify::add_page_heading}
\usage{
add_page_heading(list_of_lists)
}
\arguments{
\item{list_of_lists}{Encoded headers.}
}
\value{
Empty flextable object with headers assigned based on list_of_lists
}
\description{
Creates headers in the empty flextable object from the list of headers encoded as triplets/pairs/singles each
based on alignment: singles are centered, doubles are aligned to l/r corners and triplets are l/c/r divided
}
\examples{
add_page_heading(c(c("left aligned", "centered", "right aligned"),
                   c("left aligned", "right aligned")))
add_page_heading(NA)
add_page_heading(c(c("centered")))
try(add_page_heading(c(c("i", "like", "programming", 'NA'))))
}
