% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Trigger the build of the \code{clin_document}}
\usage{
build(x)
}
\arguments{
\item{x}{A \code{clin_document} object}
}
\value{
An executed \code{clin_document}
}
\description{
The functions used to assemble a \code{clin_document} object
do not trigger the processing of any data. Rather, a lazy execution style
is used to allow you to construct your table and then explicitly state when
the data processing should happen. \code{build} triggers this event.
}
\details{
When the \code{build} command is executed, all of the data
processing commences. Any pre-processing necessary within the table
environment takes place first. Next, each of the layers begins executing.
Once the layers complete executing, the output of each layer is stacked into
the resulting data frame.

Once this process is complete, any post-processing necessary within the table
environment takes place, and the final output can be delivered.
}
\examples{
# Load in Pipe
library(magrittr)

clin_document(iris) \%>\% build()

}
\seealso{
clin_document
}
