% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{clin_default_table_width}
\alias{clin_default_table_width}
\title{Get the Default Table Width for Clinical Documents}
\usage{
clin_default_table_width()
}
\value{
An rdocx object from the officer package
}
\description{
This function calculates the default table width based on the page width
and margins specified in the \code{clinify_docx_default} option.
}
\examples{
clin_default_table_width()
}
