% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{clin_document}
\alias{clin_document}
\title{Create a Clinify document object}
\usage{
clin_document(
  target,
  headers = NULL,
  titles = NULL,
  footnotes = NULL,
  maxrows = NULL,
  group_cols = NULL,
  pagination_strategy = NULL,
  title_style = header_style_default,
  table_style = table_style_default,
  footnote_style = footer_style_default
)
}
\arguments{
\item{target}{Dataset that will be formatted and saved as .docx}

\item{headers}{List of strings that will be used to name columns}

\item{titles}{List of strings that will be displayed on top of each page}

\item{footnotes}{List of strings that will be displayed at the bottom of each page}

\item{maxrows}{Maximum amount of rows that would be put on 1 page of the document}

\item{group_cols}{List of column names that indicates which rows should be displayed on the same page}

\item{pagination_strategy}{Matrix that shows how target should be split to fit on pages}

\item{title_style}{Style that will be applied to the output document`s titles}

\item{table_style}{Style that will be applied to the output document` body(table)}

\item{footnote_style}{Style that will be applied to the output document`s footnotes}
}
\value{
A \code{clin_document} object
}
\description{
The \code{clin_document} object is the main container from which  is constructed. Clinify tables are made up of
one or more pages. Document contains titles, footnotes, original data in the form of the data.frame, as well as page and table attributes.
}
\details{
When a \code{clin_document} is created, it will contain the following bindings:
\itemize{
\item{target - Dataset that will be formatted}
\item{headers - Vector containing table headers. This defaults to the target dataset's column names}
\item{titles - List of strings that will be displayed at the top of each page}
\item{footnotes - List of strings that will be displayed at the bottom of each page}
\item{maxrows - Maximum amount of rows that would be put on 1 page of the document}
\item{group_cols - List of column names that indicates which rows should be displayed on the same page}
\item{pagination_strategy - Matrix that describes how a target should be paginated both vertically and horizontally}
\item{style - A style that will be applied to the output document\verb{s titles\} \\item\{style - A style that will be applied to the output document}s body(table)}
\item{style - A style that will be applied to the output document`s footnotes}
}

\code{clin_document} allows you a basic interface to instantiate the object. Modifier functions are available to change
individual parameters.
}
\examples{

doc <- clin_document(iris)

}
