% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{clin_page}
\alias{clin_page}
\title{Create a Clinify page object}
\usage{
clin_page(
  target,
  headers = colnames(target),
  titles = NA,
  footnotes = NA,
  internal_footnotes = NA,
  group_cols = NULL,
  style = NA
)
}
\arguments{
\item{target}{Dataset that will be placed on a page}

\item{headers}{List of strings that will be used to name columns}

\item{titles}{List of strings that will be displayed on top of page}

\item{footnotes}{List of strings that will be displayed at the bottom of the page}

\item{internal_footnotes}{List of strings that will be displayed at the bottom of the table, above footnotes}

\item{style}{Style that will be applied to the page contents}
}
\value{
A \code{clin_page} object
}
\description{
The \code{clin_page} object is the building block of Clinify clin_document. Page contains titles,
footnotes, data in the form of the data.frame, as well as page and table attributes.
Upon building the page, a flextable object will be created that can then be saved as
.docx file.
}
\details{
When a \code{clin_page} is created, it will contain the following attributes:
\itemize{
\item{target - Dataset that will be formatted}
\item{headers - Vector containing table headers. This defaults to the target dataset's column names}
\item{titles - List of strings that will be displayed at the top of each page}
\item{footnotes - List of strings that will be displayed at the bottom of each page}
\item{internal_footnotes - List of strings that will be displayed at the bottom of the table, above footnotes}
\item{style - A style that will be applied to the output document}
}

\code{clin_page} allows you a basic interface to instantiate the object. Modifier functions are available to change
individual parameters.
}
\examples{

tab <- clin_page(iris)

}
