% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{clinify_docx_default}
\alias{clinify_docx_default}
\title{Function to create a default officer docx object}
\usage{
clinify_docx_default()
}
\value{
An rdocx object from the officer package
}
\description{
Default document attributes assume landscape A4 paper size
}
