% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_col_widths}
\alias{get_col_widths}
\title{The function to calculate widths of data columns given the label column(-s) width(-s).}
\usage{
get_col_widths(
  x,
  lbl_width = 2,
  num_data_cols = 3,
  page_width = 11.42,
  left_margin = 0.56,
  right_margin = 1.06
)
}
\arguments{
\item{x}{Dataframe under analysis.}

\item{lbl_width}{Width of the label column(s).  Default is c(2,)}

\item{num_data_cols}{Number of data columns in resulting table. Default is 3}

\item{page_width}{Expected page width. Default is 11.42 inch}

\item{left_margin}{Expected page left margin. Default is 0.56 inch}

\item{right_margin}{Expected page right margin. Default is 1.06}
}
\value{
Size of resulting data columns (all equal) in inches.
}
\description{
get_col_widths(df, lbl_width = 4, num_data_cols = 4) - this will calculate the widths for each of four data columns (equally), given the label column has 4 in width.
}
\examples{
get_col_widths(df, lbl_width = 4, num_data_cols = 4)
}
