% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_group_starts_from_df}
\alias{get_group_starts_from_df}
\title{The function to get all group-starting rows from rowlabel column}
\usage{
get_group_starts_from_df(df, row_label_no = 1)
}
\arguments{
\item{df}{Dataframe under analysis.}

\item{row_label_no}{Order number of rowlabel column.}
}
\value{
Vector of the strings that are group-starting rows from rowlabel column
}
\description{
The function to get all group-starting rows from rowlabel column
}
