% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_groups_from_df}
\alias{get_groups_from_df}
\title{The function to get row numbers of same page groups}
\usage{
get_groups_from_df(df, row_label_no = 1)
}
\arguments{
\item{df}{Dataframe under analysis.}

\item{row_label_no}{Order number of rowlabel column.}
}
\value{
Vector of the row numbers of groups that should be kept on the same page
}
\description{
The function to get row numbers of groups that should be kept on the same page
if we need rows 2-5 to be on the same page, this function will return c(2,3,4)
'keep_with_next' attribute will be applied to those rows in Word.
Currently, groups are determined by a tab. If next row has a tab in rowlbl,
then it is considered part of the current group. When a row without tab is encountered,
the group starts anew.
}
