% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pagination_method}
\alias{get_pagination_method}
\title{Get the pagination method of a clintable object}
\usage{
get_pagination_method(x)
}
\arguments{
\item{x}{A clintable object}
}
\value{
A character string specifying the pagination method
}
\description{
Get the pagination method of a clintable object
}
\examples{

ct <- clintable(mtcars)

get_pagination_method(ct)

}
