% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{paginate}
\alias{paginate}
\title{Process source data to create separate pages}
\usage{
paginate(x, ...)
}
\arguments{
\item{x}{a clin_document object}

\item{...}{arguments passed to dispatch}
}
\value{
The clin_document object with 'pages' attribute
}
\description{
This is an internal method, but is exported to support S3 dispatch. Not intended for direct use by a user.
}
\keyword{internal}
