% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_alternating}
\alias{print_alternating}
\title{Method for printing alternating pages}
\usage{
print_alternating(x, n)
}
\arguments{
\item{x}{a clintable object}

\item{n}{number of pages within the clintable to print}
}
\description{
Method for printing alternating pages
}
