% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_with_pagenums}
\alias{replace_with_pagenums}
\title{Replace Table Cells with Word Page Number Fields}
\usage{
replace_with_pagenums(x)
}
\arguments{
\item{x}{A clintable object}
}
\value{
A clintable object
}
\description{
This helper function will find placeholder text and replace the fields of
the flextable object with the appropriate page number fields. This
allows you to current and total page fields within Word documents.
Note that this is intended to be used in the defaults for
clinify_titles_default or clinify_footnotes_default.
}
\examples{

title <- new_title_footnote(
  list(
    # We'll add tools to automate paging
    c("Protocol: CDISCPILOT01", "Page X of Y"),
    c("Table 14-2.01"),
    c("Summary of Demographic and Baseline Characteristics")
  ),
  "titles"
)
  
title <- replace_with_pagenums(title)

}
