% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save}
\alias{save}
\title{save the \code{clin_document} to specified location}
\usage{
save(x, ...)
}
\arguments{
\item{x}{A \code{clin_document} object

\item{target - a full path to the document}}
}
\description{
\code{save} saves the \code{clin_document} and returns it invisibly.
}
\details{
By default, \code{save} method saves the document to the current folder
with the name of 'my_doc.docx'
}
\examples{
# Load in Pipe
library(magrittr)

doc <- clin_document(iris) \%>\% build()
save(doc)

}
\seealso{
clin_document
}
