% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_clintable.R
\name{slice_clintable}
\alias{slice_clintable}
\title{Slice a clintable object}
\usage{
slice_clintable(x, rows, columns)
}
\arguments{
\item{x}{A flextable object}

\item{rows}{Subset of rows to extract}

\item{columns}{Subset of columns to extract}
}
\value{
A clintable object
}
\description{
This function takes in a clintable (or a flextable) object and slices the
columns based on the provided rows and columns, allowing you to effectively
subset a flextable object that has already had styling applied, chunking it
as desired.
}
\examples{
x <- flextable(mtcars)
y <- slice_clintable(x, 20:32, c(1:3, 5:8))
z <- flextable(mtcars[20:32, c(1:3, 5:8)])
}
