% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{table_style_default}
\alias{table_style_default}
\title{The function to apply default table styling}
\usage{
table_style_default(x, ...)
}
\arguments{
\item{x}{flextable object that is modified by the function.}

\item{...}{any additional arguments}
}
\value{
None
}
\description{
Headers with the same value are merged both horizontally and vertically
All borders are clear except for the header (bleck, solid, 0.2pt)
As well as top horizontal line for the table header.
font style - Courier New, 9pt, not bold, not italic
padding for table body - bottom = 0.1, top = 0.1
padding for table first header - top = 9
padding for table last header - bottom = 9
table's layout is set to be fixed.
}
