% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_adas}
\alias{test_adas}
\title{ADAS Data}
\format{
A data.frame with 1,040 rows and 40 columns.
}
\source{
https://github.com/phuse-org/TestDataFactory
}
\usage{
test_adas
}
\description{
A subset of the PHUSE Test Data Factory ADAS data set.
}
\seealso{
\code{\link[=get_data_labels]{get_data_labels()}}
}
\keyword{datasets}
