% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{write_clintable}
\alias{write_clintable}
\title{Clintable print method}
\usage{
write_clintable(x, file, apply_defaults = TRUE, settings = NULL)
}
\arguments{
\item{x}{a clintable object}

\item{file}{The file path to which the file should be written}

\item{apply_defaults}{Apply default styles. These styles are stored in the
options clinify_header_default, clinify_footer_default, and
clinify_table_default respectively. Defaults to true.}

\item{settings}{Settings that will be passed}
}
\value{
Invisible
}
\description{
Extraction of flextable print method with special handling of clintable pages
and
}
\examples{

ct <- clintable(mtcars)

print(ct)

ct <- clin_alt_pages(
  ct,
  key_cols = c('mpg', 'cyl', 'hp'),
  col_groups = list(
    c('disp', 'drat', 'wt'),
    c('qsec', 'vs', 'am'),
    c('gear', 'carb')
  ),
  max_rows = 10
)

print(ct)

}
