% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_page_footer.R
\name{add_page_footer}
\alias{add_page_footer}
\title{FUNCTION to add page footer row.}
\usage{
add_page_footer(list_of_lists)
}
\arguments{
\item{list_of_lists}{Encoded footers.}
}
\value{
Empty flextable object with footers assigned based on list_of_lists
}
\description{
Creates footers in the empty flextable object from the list of footers encoded as triplets/pairs/singles each
based on alignment: singles are centered, doubles are aligned to l/r corners and triplets are l/c/r divided
}
\examples{
add_page_footer(c(c("left aligned", "centered", "right aligned"),
                  c("left aligned", "right aligned")))
add_page_footer(NA)
add_page_footer(c(c("centered")))
try(add_page_footer(c(c("i", "like", "programming", 'NA'))))
}
