% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clintable.R
\name{as_clintable}
\alias{as_clintable}
\title{Convert a flextable into a clintable object}
\usage{
as_clintable(ft)
}
\arguments{
\item{ft}{A flextable object}
}
\value{
A clintable object
}
\description{
Convert a flextable into a clintable object
}
\examples{
ft <- flextable(mtcars)
as_clintable(ft)
}
