% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{clin_default_table_width}
\alias{clin_default_table_width}
\title{Helper function to get the total page width}
\usage{
clin_default_table_width()
}
\value{
An rdocx object from the officer package
}
\description{
Using the document dimensions stored in options, calculate the
width that can be used in table body, titles, and footnotes. Calculation
uses the page width subtracting right and left margins.
}
