% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{clin_page_by}
\alias{clin_page_by}
\title{Configure pagination using a page variable}
\usage{
clin_page_by(x, page_by, max_rows = 10)
}
\arguments{
\item{x}{A clintable object}

\item{byvar}{A column in the table used for page grouping}
}
\value{
A clintable object
}
\description{
Configure pagination using a page variable
}
