% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clindoc.R
\name{clindoc}
\alias{clindoc}
\alias{as_clindoc}
\title{Convert a Clintable to a Word Document}
\usage{
clindoc(x, ...)

as_clindoc(x, apply_defaults = TRUE)
}
\arguments{
\item{x}{A \code{clintable} object to be converted.}

\item{...}{Additional arguments (currently unused).}

\item{apply_defaults}{Logical, whether to apply default title, footnote,
and table formatting. Defaults to \code{TRUE}.}
}
\value{
An \code{officer::rdocx} object representing the formatted Word document.
}
\description{
These functions convert a \code{clintable} object into a Word document using
officer and flextable formatting. They apply default styles for titles,
footnotes, and table settings while supporting pagination options.
}
\details{
\itemize{
\item \code{clindoc()} ensures that the input is a \code{clintable} and then converts it
to a Word document.
\item \code{as_clindoc()} performs the conversion, applying default formatting and
handling pagination if specified.
}
}
\examples{

ct <- clintable(mtcars)

clindoc(ct)


}
