% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinpage.R
\name{clintable}
\alias{clintable}
\title{Create a new clintable object}
\usage{
clintable(data, ...)
}
\arguments{
\item{data}{A data frame}

\item{...}{Parameters to pass to \code{flextable::flextable()}}
}
\value{
A clintable object
}
\description{
A clintable object directly inherits from a flextable object. This function
will pass all necessary parameters \code{flextable::flextable()} and conver the
object to a \code{clintable}
}
\examples{
clintable(mtcars)
}
