% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clintable.R
\name{clintable}
\alias{clintable}
\title{Create a new clintable object}
\usage{
clintable(x, page_by = NULL, group_by = NULL, ...)
}
\arguments{
\item{x}{A data frame}

\item{page_by}{A variable in the input dataframe to use for pagination}

\item{group_by}{A variable which will be used for grouping and attached
as a label above the table headers}

\item{...}{Parameters to pass to \code{flextable::flextable()}}
}
\value{
A clintable object
}
\description{
A clintable object directly inherits from a flextable object. This function
will pass all necessary parameters \code{flextable::flextable()} and conver the
object to a \code{clintable}
}
\examples{
clintable(mtcars)
}
