% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_titles_footnotes.R
\name{new_title_footnote}
\alias{new_title_footnote}
\title{Create a new title or footnote flextable}
\usage{
new_title_footnote(x, sect = c("titles", "footnotes"))
}
\arguments{
\item{x}{a list of character vectors, no more than 3 elements to a vector.}

\item{sect}{Either "titles" or "footnotes"}
}
\value{
A flextable object
}
\description{
Create a new title or footnote flextable
}
\examples{
#TODO:
}
