% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.clintable}
\alias{print.clintable}
\title{Clintable print method}
\usage{
\method{print}{clintable}(x, n = 3, ...)
}
\arguments{
\item{x}{a clintable object}

\item{n}{number of pages within the clintable to print}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\value{
Invisible
}
\description{
Extraction of flextable print method with special handling of clintable pages
and
}
