% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save}
\alias{save}
\title{save the \code{clin_document} to specified location}
\usage{
save(x)
}
\arguments{
\item{x}{A \code{clin_document} object}
}
\description{
save the \code{clin_document} to specified location
}
\examples{
# Load in Pipe
library(magrittr)

doc <- clin_document(iris) \%>\% build()
save(doc)

}
\seealso{
clin_document
}
