% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{write_clintable}
\alias{write_clintable}
\title{Clintable print method}
\usage{
write_clintable(x, file, settings = getOption("clin_doc_settings"), ...)
}
\arguments{
\item{x}{a clintable object}

\item{...}{Additional parameters passed to flextable print method}
}
\value{
Invisible
}
\description{
Extraction of flextable print method with special handling of clintable pages
and
}
\examples{

ct <- clintable(mtcars)

print(ct)

ct <- clin_alt_pages(
  ct,
  key_cols = c('mpg', 'cyl', 'hp'),
  col_groups = list(
    c('disp', 'drat', 'wt'),
    c('qsec', 'vs', 'am'),
    c('gear', 'carb')
  ),
  max_rows = 10
)

print(ct)

}
